separated instanceVolume attach modals and actions
